Fix release 2.5.0.011
git-svn-id: https://plugins.svn.wordpress.org/wp-photo-album-plus/trunk@337816 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ wppa_upload.php
@@ -409,7 +409,7 @@ global $wpdb;
 							$pcount++;
 							if ($delp) {
 								unlink($file);
-								unlink($meta);
+								if (is_file($meta)) unlink($meta);
 							}
 						}
 						else {
